% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threads.R
\name{accrqa_get_max_threads}
\alias{accrqa_get_max_threads}
\title{Get maximum available threads (OpenMP)}
\usage{
accrqa_get_max_threads()
}
\value{
integer
}
\description{
Get maximum available threads (OpenMP)
}
