% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAccRQA.R
\name{accrqa_RR_target}
\alias{accrqa_RR_target}
\title{Find threshold that yields target RR (recurrence rate)}
\usage{
accrqa_RR_target(
  input_data,
  tau,
  emb,
  target_RR,
  epsilon = 0.01,
  distance_type = "euclidean",
  comp_platform = "cpu",
  max_iter = 20,
  threshold_min = 0,
  threshold_max = NULL,
  n_threshold = 10,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{input_data}{Numeric vector (time series)}

\item{tau}{Delay}

\item{emb}{Embedding dimension}

\item{target_RR}{Target recurrence(s) rate in \code{[0, 1]}. Can be scalar or vector.}

\item{epsilon}{OPTIONAL Allowed error to find the desired RR. Default: 0.01}

\item{distance_type}{OPTIONAL Distance type ("euclidean" or "maximal"). Default: "euclidean".}

\item{comp_platform}{OPTIONAL Computer platform to use either "cpu" or "nv_gpu". Default: "cpu".}

\item{max_iter}{OPTIONAL Maximum iterations to find the threshold before give up. Default: 20.}

\item{threshold_min}{OPTIONAL Lower bound for search. Default: 0}

\item{threshold_max}{OPTIONAL Upper bound for search (set automatically if NULL)}

\item{n_threshold}{OPTIONAL How many thresholds per iteration (e.g. 10). Default: 10.}

\item{verbose}{OPTIONAL If TRUE, prints progress. Default: FALSE.}

\item{...}{Further arguments (currently ignored but kept for S3 compatibility).}
}
\value{
Data frame with columns:
\itemize{
\item target_RR - target RR
\item threshold - threshold to achieve desired RR
\item RR_found  - RR at the found threshold
}
}
\description{
Find threshold that yields target RR (recurrence rate)
}
\examples{
x <- runif(1000)
accrqa_RR_target(x, tau = 1, emb = 2, target_RR = 0.5, epsilon = 0.001)

# multiple targets
accrqa_RR_target(x, tau = 1, emb = 2, target_RR = c(0.1,0.5,0.8), epsilon = 0.001)
}
