% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threads.R
\name{accrqa_set_num_threads}
\alias{accrqa_set_num_threads}
\title{Set number of threads used by AccRQA (OpenMP)}
\usage{
accrqa_set_num_threads(n)
}
\arguments{
\item{n}{integer >= 1}
}
\value{
invisible(NULL)
}
\description{
Set number of threads used by AccRQA (OpenMP)
}
