% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAccRQA.R
\name{plot.accrqa_rp}
\alias{plot.accrqa_rp}
\title{Plot method for AccRQA recurrence plot objects}
\usage{
\method{plot}{accrqa_rp}(
  x,
  title = "",
  xlabel = "Time",
  ylabel = "Time",
  style = c("raster", "tile", "point"),
  color_min = "#eff3ff",
  color_max = "#08519c",
  x_dates = NULL,
  y_dates = x_dates,
  summary = FALSE,
  ...
)
}
\arguments{
\item{x}{An object of class "accrqa_rp".}

\item{title}{Plot title.}

\item{xlabel}{X-axis label.}

\item{ylabel}{Y-axis label.}

\item{style}{One of "raster", "tile", or "point".}

\item{color_min}{Colour for 0 entries.}

\item{color_max}{Colour for 1 entries.}

\item{x_dates}{Optional vector of dates/times for the x-axis (length = RP size).}

\item{y_dates}{Optional vector of dates/times for the y-axis (length = RP size).}

\item{summary}{Optional boolean parameter for adding metadata to the right side of the plot.}

\item{...}{Further arguments (currently ignored but kept for S3 compatibility).}
}
\description{
Plot method for AccRQA recurrence plot objects
}
\examples{
x <- seq(0, 10*pi, length.out = 36)
ts <- sin(x)
rp <- accrqa_RP(ts, tau = 1, emb = 2, threshold = 0.1, distance_type = "maximal")
# raster plot
plot(rp)

# tile style
plot(rp, style = "tile")

# With dates:
time_index <- as.Date("2020-01-01") + 0:(sqrt(length(rp$output)) - 1)
plot(rp, x_dates = time_index, y_dates = time_index)

plot(rp, x_dates = time_index, y_dates = time_index, summary = TRUE)
}
